+++ text-provider/src/test/java/edu/kit/kastel/mcse/ardoco/core/text/providers/corenlp/CoreNLPProviderTest.java
@@ -4,22 +4,17 @@ package edu.kit.kastel.mcse.ardoco.core.text.providers.corenlp;
 import java.io.FileInputStream;
 import java.io.FileNotFoundException;
 
-import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.BeforeAll;
-import org.junit.jupiter.api.Test;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-class CoreNLPProviderTest {
+import edu.kit.kastel.mcse.ardoco.core.text.providers.ITextConnector;
+import edu.kit.kastel.mcse.ardoco.core.text.providers.base.ProviderTest;
+
+class CoreNLPProviderTest extends ProviderTest {
     private static final Logger logger = LoggerFactory.getLogger(CoreNLPProviderTest.class);
     protected static String inputText = "src/test/resources/teastore.txt";
 
     private static CoreNLPProvider coreNLPProvider = null;
-
-    @BeforeAll
-    static void beforeAll() {
-        coreNLPProvider = getCoreNLPProvider();
-    }
 
     public synchronized static CoreNLPProvider getCoreNLPProvider() {
         if (coreNLPProvider == null) {
@@ -32,9 +27,8 @@ class CoreNLPProviderTest {
         return coreNLPProvider;
     }
 
-    @Test
-    void getTextTest() {
-        var text = coreNLPProvider.getAnnotatedText();
-        Assertions.assertNotNull(text);
+    @Override
+    protected ITextConnector getProvider() {
+        return CoreNLPProviderTest.getCoreNLPProvider();
     }
 }
