+++ text-provider/src/test/java/edu/kit/kastel/mcse/ardoco/core/text/providers/json/JsonSentenceTest.java
@@ -2,10 +2,9 @@
 package edu.kit.kastel.mcse.ardoco.core.text.providers.json;
 
 import edu.kit.kastel.mcse.ardoco.core.text.providers.ITextConnector;
-import edu.kit.kastel.mcse.ardoco.core.text.providers.base.PhraseTest;
+import edu.kit.kastel.mcse.ardoco.core.text.providers.base.SentenceTest;
 
-class JsonPhraseTest extends PhraseTest {
-
+class JsonSentenceTest extends SentenceTest {
     @Override
     protected ITextConnector getProvider() {
         return JsonTextProviderTest.getTextProvider();
